Create classes.html page template
The class detail page is getting specific enough that it makes sense
to break it out into its own template now. The conditional variables
and markup I’ve used up until now were okay for early prototyping,
but are getting too complex to hold in a single template.

This commit also updates the page generator settings in _config.yml
and sets the page.html template back to the framework default for
the other content types that still use it (directory & feed).

Add [share] and [register] buttons with basic Foundation classes.