Change data labels to reflect content
The generic _data.xmls content categories worked as a starting
point, but as the page templates are starting to get more complex,
it will be easier in the long run to make these labels more
semantically correct so that they are easier to recognize.

This commit also updates the classes.html page template variables
and renames categories.md to classes.md.