Create global nav drawer
Update Navigation sheet of _data file to reflect a category
structure in the drawer nav. Responsive drawer nav pattern is
moved from an include to the default page for clarity – there
is a whole-page wrap that occurs that is hard to follow if
opening and closing divs appear in different includes.

This commit also removes CSS that applies only to the original
nav-bar pattern as well as the original nav-bar include.